Refactor database client source
