+++ stdlib/jdbc/src/main/java/org/ballerinax/jdbc/actions/Call.java
@@ -44,4 +44,8 @@ public class Call {
                 parameters, strand);
         return callStatement.execute();
     }
+
+    private Call() {
+
+    }
 }
