+++ stdlib/jdbc/src/main/java/org/ballerinax/jdbc/functions/Close.java
@@ -22,7 +22,7 @@ import org.ballerinalang.jvm.values.ObjectValue;
 import org.ballerinalang.natives.annotations.BallerinaFunction;
 import org.ballerinax.jdbc.Constants;
 import org.ballerinax.jdbc.datasource.SQLDatasource;
-import org.ballerinax.jdbc.datasource.SQLDatasourceUtils;
+import org.ballerinax.jdbc.exceptions.ErrorGenerator;
 
 /**
  * {@code Close} is the Close function implementation of the JDBC client connection pool.
@@ -44,7 +44,7 @@ public class Close {
             try {
                 datasource.decrementClientCounterAndAttemptPoolShutdown();
             } catch (InterruptedException e) {
-                return  SQLDatasourceUtils.getSQLApplicationError("Error while stopping the database client");
+                return  ErrorGenerator.getSQLApplicationError("Error while stopping the database client");
             }
         }
         return null;
