+++ stdlib/jdbc/src/main/java/org/ballerinax/jdbc/functions/CreateClient.java
@@ -17,13 +17,16 @@
  */
 package org.ballerinax.jdbc.functions;
 
+import org.ballerinalang.jvm.BallerinaValues;
 import org.ballerinalang.jvm.Strand;
 import org.ballerinalang.jvm.values.MapValue;
 import org.ballerinalang.jvm.values.ObjectValue;
 import org.ballerinalang.natives.annotations.BallerinaFunction;
 import org.ballerinax.jdbc.Constants;
-import org.ballerinax.jdbc.datasource.SQLDatasourceUtils;
+import org.ballerinax.jdbc.datasource.PoolOptionsWrapper;
+import org.ballerinax.jdbc.datasource.SQLDatasource;
 
+import java.util.Locale;
 import java.util.UUID;
 
 /**
@@ -41,8 +44,44 @@ public class CreateClient {
 
     public static ObjectValue createClient(Strand strand, MapValue<String, Object> config, MapValue<String,
                 Object> globalPoolOptions) {
-        ObjectValue jdbcClient = SQLDatasourceUtils.createSQLDBClient(config, globalPoolOptions);
+        ObjectValue jdbcClient = createSQLDBClient(config, globalPoolOptions);
         jdbcClient.addNativeData(Constants.CONNECTOR_ID_KEY, UUID.randomUUID().toString());
         return jdbcClient;
+    }
+
+    public static ObjectValue createSQLDBClient(MapValue<String, Object> clientEndpointConfig,
+            MapValue<String, Object> globalPoolOptions) {
+        String url = clientEndpointConfig.getStringValue(Constants.EndpointConfig.URL);
+        String username = clientEndpointConfig.getStringValue(Constants.EndpointConfig.USERNAME);
+        String password = clientEndpointConfig.getStringValue(Constants.EndpointConfig.PASSWORD);
+        MapValue<String, Object> dbOptions = (MapValue<String, Object>) clientEndpointConfig
+                .getMapValue(Constants.EndpointConfig.DB_OPTIONS);
+        MapValue<String, Object> poolOptions = (MapValue<String, Object>) clientEndpointConfig
+                .getMapValue(Constants.EndpointConfig.POOL_OPTIONS);
+        boolean userProvidedPoolOptionsNotPresent = poolOptions == null;
+        if (userProvidedPoolOptionsNotPresent) {
+            poolOptions = globalPoolOptions;
+        }
+        PoolOptionsWrapper poolOptionsWrapper = new PoolOptionsWrapper(poolOptions);
+        String dbType = url.split(":")[1].toUpperCase(Locale.getDefault());
+
+        SQLDatasource.SQLDatasourceParamsBuilder builder = new SQLDatasource.SQLDatasourceParamsBuilder(dbType);
+        SQLDatasource.SQLDatasourceParams sqlDatasourceParams = builder.withPoolOptions(poolOptionsWrapper)
+                .withJdbcUrl(url).withUsername(username).withPassword(password).withDbOptionsMap(dbOptions)
+                .withIsGlobalDatasource(userProvidedPoolOptionsNotPresent).build();
+
+        return createSQLClient(sqlDatasourceParams);
+    }
+
+    private static ObjectValue createSQLClient(SQLDatasource.SQLDatasourceParams sqlDatasourceParams) {
+        SQLDatasource sqlDatasource = sqlDatasourceParams.getPoolOptionsWrapper()
+                .retrieveDatasource(sqlDatasourceParams);
+        ObjectValue sqlClient = BallerinaValues.createObjectValue(Constants.JDBC_PACKAGE_PATH, Constants.JDBC_CLIENT);
+        sqlClient.addNativeData(Constants.JDBC_CLIENT, sqlDatasource);
+        return sqlClient;
+    }
+
+    private CreateClient() {
+
     }
 }
