+++ stdlib/jdbc/src/main/java/org/ballerinax/jdbc/statement/BatchUpdateStatement.java
@@ -27,9 +27,9 @@ import org.ballerinalang.jvm.values.freeze.State;
 import org.ballerinalang.jvm.values.freeze.Status;
 import org.ballerinax.jdbc.Constants;
 import org.ballerinax.jdbc.datasource.SQLDatasource;
-import org.ballerinax.jdbc.datasource.SQLDatasourceUtils;
 import org.ballerinax.jdbc.exceptions.ApplicationException;
 import org.ballerinax.jdbc.exceptions.DatabaseException;
+import org.ballerinax.jdbc.exceptions.ErrorGenerator;
 
 import java.sql.BatchUpdateException;
 import java.sql.Connection;
@@ -86,7 +86,9 @@ public class BatchUpdateStatement extends AbstractSQLStatement {
             for (int index = 0; index < paramArrayCount; index++) {
                 ArrayValue params = (ArrayValue) parameters.getValue(index);
                 ArrayValue generatedParams = constructParameters(params);
-                createProcessedStatement(conn, stmt, generatedParams, datasource.getDatabaseProductName());
+                ProcessedStatement processedStatement = new ProcessedStatement(conn, stmt, generatedParams,
+                        datasource.getDatabaseProductName());
+                stmt = processedStatement.prepare();
                 stmt.addBatch();
             }
             updatedCount = stmt.executeBatch();
@@ -112,22 +114,22 @@ public class BatchUpdateStatement extends AbstractSQLStatement {
             }
             handleErrorOnTransaction(this.strand);
             return createFrozenBatchUpdateResultRecord(createUpdatedCountArray(updatedCount, paramArrayCount),
-                    SQLDatasourceUtils.getSQLDatabaseError(e, errorMessagePrefix + ": "));
+                    ErrorGenerator.getSQLDatabaseError(e, errorMessagePrefix + ": "));
         } catch (SQLException e) {
              handleErrorOnTransaction(this.strand);
              //checkAndObserveSQLError(context, e.getMessage());
             return createFrozenBatchUpdateResultRecord(createUpdatedCountArray(null, paramArrayCount),
-                    SQLDatasourceUtils.getSQLDatabaseError(e, errorMessagePrefix + ": "));
+                    ErrorGenerator.getSQLDatabaseError(e, errorMessagePrefix + ": "));
         } catch (DatabaseException e) {
             handleErrorOnTransaction(this.strand);
             // checkAndObserveSQLError(context, e.getMessage());
             return createFrozenBatchUpdateResultRecord(createUpdatedCountArray(null, paramArrayCount),
-                    SQLDatasourceUtils.getSQLDatabaseError(e, errorMessagePrefix + ": "));
+                    ErrorGenerator.getSQLDatabaseError(e, errorMessagePrefix + ": "));
         } catch (ApplicationException e) {
             handleErrorOnTransaction(this.strand);
             // checkAndObserveSQLError(context, e.getMessage());
             return createFrozenBatchUpdateResultRecord(createUpdatedCountArray(null, paramArrayCount),
-                    SQLDatasourceUtils.getSQLApplicationError(e, errorMessagePrefix + ": "));
+                    ErrorGenerator.getSQLApplicationError(e, errorMessagePrefix + ": "));
         } finally {
             cleanupResources(stmt, conn, !isInTransaction);
         }
