+++ stdlib/jdbc/src/main/java/org/ballerinax/jdbc/statement/CallStatement.java
@@ -34,6 +34,7 @@ import org.ballerinax.jdbc.datasource.SQLDatasource;
 import org.ballerinax.jdbc.datasource.SQLDatasourceUtils;
 import org.ballerinax.jdbc.exceptions.ApplicationException;
 import org.ballerinax.jdbc.exceptions.DatabaseException;
+import org.ballerinax.jdbc.exceptions.ErrorGenerator;
 
 import java.io.IOException;
 import java.math.BigDecimal;
@@ -93,7 +94,9 @@ public class CallStatement extends AbstractSQLStatement {
             ArrayValue generatedParams = constructParameters(parameters);
             conn = getDatabaseConnection(strand, client, datasource, false);
             stmt = getPreparedCall(conn, datasource, query, generatedParams);
-            createProcessedStatement(conn, stmt, generatedParams, datasource.getDatabaseProductName());
+            ProcessedStatement processedStatement = new ProcessedStatement(conn, stmt, generatedParams,
+                    datasource.getDatabaseProductName());
+            stmt = (CallableStatement) processedStatement.prepare();
             boolean refCursorOutParamsPresent = isRefCursorOutParamPresent(generatedParams);
             boolean resultSetsReturned = false;
             TableResourceManager rm = null;
@@ -124,17 +127,17 @@ public class CallStatement extends AbstractSQLStatement {
             cleanupResources(resultSets, stmt, conn, !isInTransaction);
             handleErrorOnTransaction(this.strand);
             // checkAndObserveSQLError(context, "execute stored procedure failed: " + e.getMessage());
-            return SQLDatasourceUtils.getSQLDatabaseError(e, errorMessagePrefix);
+            return ErrorGenerator.getSQLDatabaseError(e, errorMessagePrefix);
         } catch (DatabaseException e) {
             cleanupResources(resultSets, stmt, conn, !isInTransaction);
             handleErrorOnTransaction(this.strand);
             // checkAndObserveSQLError(context, "execute stored procedure failed: " + e.getMessage());
-            return SQLDatasourceUtils.getSQLDatabaseError(e, errorMessagePrefix);
+            return ErrorGenerator.getSQLDatabaseError(e, errorMessagePrefix);
         } catch (ApplicationException e) {
             cleanupResources(resultSets, stmt, conn, !isInTransaction);
             handleErrorOnTransaction(this.strand);
             // checkAndObserveSQLError(context, "execute stored procedure failed: " + e.getMessage());
-            return SQLDatasourceUtils.getSQLApplicationError(e, errorMessagePrefix);
+            return ErrorGenerator.getSQLApplicationError(e, errorMessagePrefix);
         }
         return null;
     }
@@ -216,8 +219,8 @@ public class CallStatement extends AbstractSQLStatement {
             }
             MapValue<String, Object> paramValue = (MapValue<String, Object>) params.getValue(index);
             if (paramValue != null) {
-                String sqlType = getSQLType(paramValue);
-                int direction = getParameterDirection(paramValue);
+                String sqlType = StatementProcessUtils.getSQLType(paramValue);
+                int direction = StatementProcessUtils.getParameterDirection(paramValue);
                 if (direction == Constants.QueryParamDirection.INOUT
                         || direction == Constants.QueryParamDirection.OUT) {
                     setOutParameterValue(stmt, sqlType, index, paramValue, rm);
@@ -396,7 +399,7 @@ public class CallStatement extends AbstractSQLStatement {
         int paramCount = params.size();
         for (int index = 0; index < paramCount; index++) {
             MapValue<String, Object> paramValue = (MapValue<String, Object>) params.getRefValue(index);
-            int direction = getParameterDirection(paramValue);
+            int direction = StatementProcessUtils.getParameterDirection(paramValue);
             if (direction == Constants.QueryParamDirection.OUT || direction == Constants.QueryParamDirection.INOUT) {
                 return true;
             }
@@ -410,8 +413,8 @@ public class CallStatement extends AbstractSQLStatement {
         for (int index = 0; index < paramCount; index++) {
             MapValue<String, Object> paramValue = (MapValue<String, Object>) params.getRefValue(index);
             if (paramValue != null) {
-                String sqlType = getSQLType(paramValue);
-                int direction = getParameterDirection(paramValue);
+                String sqlType = StatementProcessUtils.getSQLType(paramValue);
+                int direction = StatementProcessUtils.getParameterDirection(paramValue);
                 if (direction == Constants.QueryParamDirection.OUT && Constants.SQLDataTypes.REFCURSOR
                         .equals(sqlType)) {
                     refCursorOutParamPresent = true;
@@ -444,7 +447,7 @@ public class CallStatement extends AbstractSQLStatement {
             }
             cleanupResources(stmt, conn, connectionClosable);
         } catch (SQLException e) {
-            throw SQLDatasourceUtils.getSQLDatabaseError(e, "error in cleaning sql resources: ");
+            throw ErrorGenerator.getSQLDatabaseError(e, "error in cleaning sql resources: ");
         }
     }
 }
