+++ stdlib/jdbc/src/main/java/org/ballerinax/jdbc/statement/SelectStatement.java
@@ -26,9 +26,9 @@ import org.ballerinalang.jvm.values.ObjectValue;
 import org.ballerinalang.jvm.values.TypedescValue;
 import org.ballerinax.jdbc.Constants;
 import org.ballerinax.jdbc.datasource.SQLDatasource;
-import org.ballerinax.jdbc.datasource.SQLDatasourceUtils;
 import org.ballerinax.jdbc.exceptions.ApplicationException;
 import org.ballerinax.jdbc.exceptions.DatabaseException;
+import org.ballerinax.jdbc.exceptions.ErrorGenerator;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -73,7 +73,9 @@ public class SelectStatement extends AbstractSQLStatement {
             conn = getDatabaseConnection(strand, client, datasource, true);
             String processedQuery = createProcessedQueryString(query, generatedParams);
             stmt = getPreparedStatement(conn, datasource, processedQuery);
-            createProcessedStatement(conn, stmt, generatedParams, datasource.getDatabaseProductName());
+            ProcessedStatement processedStatement = new ProcessedStatement(conn, stmt, generatedParams,
+                    datasource.getDatabaseProductName());
+            stmt = processedStatement.prepare();
             rs = stmt.executeQuery();
             TableResourceManager rm = new TableResourceManager(conn, stmt, true);
             List<ColumnDefinition> columnDefinitions = getColumnDefinitions(rs);
@@ -84,19 +86,19 @@ public class SelectStatement extends AbstractSQLStatement {
             handleErrorOnTransaction(this.strand);
             //TODO: JBalMigration Commenting out transaction handling and observability
             // checkAndObserveSQLError(context, "execute query failed: " + e.getMessage());
-            return SQLDatasourceUtils.getSQLDatabaseError(e, errorMessagePrefix);
+            return ErrorGenerator.getSQLDatabaseError(e, errorMessagePrefix);
         } catch (DatabaseException e) {
             cleanupResources(null, stmt, conn, true);
             //TODO: JBalMigration Commenting out transaction handling and observability
             handleErrorOnTransaction(this.strand);
             // checkAndObserveSQLError(context, "execute query failed: " + e.getMessage());
-            return SQLDatasourceUtils.getSQLDatabaseError(e, errorMessagePrefix);
+            return ErrorGenerator.getSQLDatabaseError(e, errorMessagePrefix);
         } catch (ApplicationException e) {
             cleanupResources(null, stmt, conn, true);
             //TODO: JBalMigration Commenting out transaction handling and observability
             handleErrorOnTransaction(this.strand);
             // checkAndObserveSQLError(context, "execute query failed: " + e.getMessage());
-            return SQLDatasourceUtils.getSQLApplicationError(e, errorMessagePrefix);
+            return ErrorGenerator.getSQLApplicationError(e, errorMessagePrefix);
         }
     }
