+++ stdlib/jdbc/src/main/java/org/ballerinax/jdbc/table/SQLDataIterator.java
@@ -36,6 +36,7 @@ import org.ballerinalang.jvm.values.TableIterator;
 import org.ballerinalang.stdlib.time.util.TimeUtils;
 import org.ballerinax.jdbc.Constants;
 import org.ballerinax.jdbc.datasource.SQLDatasourceUtils;
+import org.ballerinax.jdbc.exceptions.ErrorGenerator;
 import org.ballerinax.jdbc.exceptions.PanickingApplicationException;
 
 import java.io.IOException;
@@ -85,7 +86,7 @@ public class SQLDataIterator extends TableIterator {
             resourceManager.gracefullyReleaseResources();
             rs = null;
         } catch (SQLException e) {
-            throw SQLDatasourceUtils.getSQLDatabaseError(e);
+            throw ErrorGenerator.getSQLDatabaseError(e);
         }
     }
 
@@ -99,14 +100,14 @@ public class SQLDataIterator extends TableIterator {
             Blob bValue = rs.getBlob(columnIndex);
             return rs.wasNull() ? null : SQLDatasourceUtils.getString(bValue);
         } catch (SQLException e) {
-            throw SQLDatasourceUtils.getSQLDatabaseError(e);
+            throw ErrorGenerator.getSQLDatabaseError(e);
         }
     }
 
     @Override
     public MapValue<String, Object> generateNext() {
         if (this.type == null) {
-            throw SQLDatasourceUtils
+            throw ErrorGenerator
                     .getSQLApplicationError("the expected record type is not specified in the remote function");
         }
         MapValue<String, Object> bStruct = new MapValueImpl<>(this.type);
@@ -116,7 +117,7 @@ public class SQLDataIterator extends TableIterator {
         try {
             BField[] structFields = this.type.getFields().values().toArray(new BField[0]);
             if (columnDefs.size() != structFields.length) {
-                throw SQLDatasourceUtils.getSQLApplicationError("Number of fields in the constraint type is " + (
+                throw ErrorGenerator.getSQLApplicationError("Number of fields in the constraint type is " + (
                         structFields.length > columnDefs.size() ?
                                 "greater" : "lower") + " than column count of the result set");
             }
@@ -225,18 +226,18 @@ public class SQLDataIterator extends TableIterator {
                             handleStructValue(bStruct, fieldName, structData, fieldType);
                             break;
                         default:
-                            throw SQLDatasourceUtils.getSQLApplicationError(
+                            throw ErrorGenerator.getSQLApplicationError(
                                     "unsupported sql type " + sqlType + " found for the column " + columnName
                                             + " index:" + index);
                     }
                 }
             }
         } catch (IOException | SQLException e) {
-            throw SQLDatasourceUtils.getSQLApplicationError(
+            throw ErrorGenerator.getSQLApplicationError(
                     "error in retrieving next value for column: " + columnName + ": of SQL Type: " + sqlType + ": "
                             + "at " + "index:" + index + ":" + e.getMessage());
         } catch (PanickingApplicationException e) {
-            throw SQLDatasourceUtils.getSQLApplicationError(e);
+            throw ErrorGenerator.getSQLApplicationError(e);
         }
         return bStruct;
     }
